remove `triton_kernels.routing`
